91social frontend assignment using React.Js
All the fields in the form group are validated.
Once the submit button is clicked, without reloading, the entered fields are showed in a different view in a dropdown modal which has an OK button which when pressed will refresh the form.
The login button on the top-right of the navbar will reveal a modal on click which also contains an uncontrolled form.
Headers and Footers along with use of Jumbotrons, Font-awesome icons,Bootstrap-social buttons are used for an enhanced user interface and smooth user experience